Added two functions for text formatting.
The bordered function allows text to be passed through as a string or list of strings and wrapped with a border.

The columns function takes a list of strings and formats them into columns. Parameters can determine whether the colums will be printed automatically or if a new list of formatted strings will be returned back as a variable.