The program is up and running!
Implemented the launcher- determines if there is a character active. Runs setup to load or create a character if not. Loops the main program with the active character. Implemented the main program- prints a list of rolls that can be made as well as an option to go to a character submenu, print the character sheet, or switch characters. Implemented character options- allows player to print a character sheet, level their character up by 1, reinitialize their character, reset the seed the RNG is running on if they want different luck, or delete their character completely. Implemented the dice roller function- takes the amount of die, type of die, and modifier as optional parameters. If none passed, User input will be taken for amount and type and modifier will be 0. Phew!